Cartridge: added INES2.0 implementation
- very messy to migrate and mix INES1.0 and INES2.0 :(

+++ cartridge/src/error.rs
@@ -11,7 +11,7 @@ pub enum CartridgeError {
     HeaderError,
     TooLargeFile(u64),
     ExtensionError,
-    MapperNotImplemented(u8),
+    MapperNotImplemented(u16),
     Others,
 }
